Update Collective profile prototype
