+++ components/crowdfunding-redesign/FundraiserLayout.tsx
@@ -1,35 +1,22 @@
 import React from 'react';
 import { useQuery } from '@apollo/client';
+import { ArrowRight } from 'lucide-react';
 import { useRouter } from 'next/router';
+import { FormattedMessage } from 'react-intl';
+import LiteYouTubeEmbed from 'react-lite-youtube-embed';
 
 import { API_V2_CONTEXT } from '../../lib/graphql/helpers';
-import LiteYouTubeEmbed from 'react-lite-youtube-embed';
-import MessageBoxGraphqlError from '../MessageBoxGraphqlError';
-import Page from '../Page';
 
-import { Banner } from './Banner';
-import { Footer } from './Footer';
-import Fundraiser from './Fundraiser';
-import { profilePageQuery } from './queries';
-import {
-  aggregateGoalAmounts,
-  getDefaultFundraiserValues,
-  getYouTubeIDFromUrl,
-  triggerPrototypeToast,
-} from './helpers';
-import { useModal } from '../ModalContext';
 import Link from '../Link';
-import { ProfileModal } from './ProfileModal';
-import { ArrowRight } from 'lucide-react';
-import { Progress } from '../ui/Progress';
-import FormattedMoneyAmount from '../FormattedMoneyAmount';
-import { StackedAvatars } from '../Avatar';
+import MessageBoxGraphqlError from '../MessageBoxGraphqlError';
 import { Button } from '../ui/Button';
-import { FormattedMessage } from 'react-intl';
-import { TabsList, TabsTrigger } from './DumbTabs';
-import { ThemeColor } from './ThemeColor';
-import ProfileLayout from './ProfileLayout';
+
 import { Goals } from './Goals';
+import { getDefaultFundraiserValues, getYouTubeIDFromUrl, triggerPrototypeToast } from './helpers';
+import ProfileLayout from './ProfileLayout';
+import { profilePageQuery } from './queries';
+import { TabsList, TabsTrigger } from './Tabs';
+import { ThemeColor } from './ThemeColor';
 
 export function FundraiserLayout({ children, activeTab }) {
   const router = useRouter();
@@ -40,7 +27,6 @@ export function FundraiserLayout({ children, activeTab }) {
   const account = data?.account;
   const fundraiser = getDefaultFundraiserValues(account);
   const tabRef = React.useRef();
-  const { showModal } = useModal();
 
   if (loading || !account) {
     return null;
@@ -68,15 +54,7 @@ export function FundraiserLayout({ children, activeTab }) {
     >
       <div className="flex min-h-screen flex-col justify-between bg-primary-foreground/75 antialiased">
         <div className="flex flex-1 flex-col">
-          {/* <Banner account={account} isFundraiser={true} /> */}
           <React.Fragment>
-            {/* <div className="border-b bg-background">
-        <div className="mx-auto flex h-16 max-w-screen-xl items-center justify-between px-6">
-          <Link href={`/preview/${account.parent?.slug || account.slug}`}>
-            <Avatar className="" collective={account.parent || account} />
-          </Link>
-        </div>
-      </div> */}
             <div className="">
               <div className="mx-auto max-w-screen-xl space-y-8 px-6 py-12">
                 <div className="space-y-4 text-center">
@@ -101,14 +79,10 @@ export function FundraiserLayout({ children, activeTab }) {
                     </div>
 
                     <p>
-                      {account.type === 'EVENT' ? 'Event' : 'Fundraiser'} by{' '}
+                      {account.type === 'EVENT' ? 'Event' : 'Project'} by{' '}
                       <Link
                         className="font-semibold text-primary hover:underline"
                         href={`/preview/${mainAccount.slug}`}
-                        onClick={e => {
-                          e.preventDefault();
-                          showModal(ProfileModal, { account: mainAccount });
-                        }}
                       >
                         {mainAccount.name} <ArrowRight className="inline align-middle" size={16} />
                       </Link>
@@ -130,37 +104,23 @@ export function FundraiserLayout({ children, activeTab }) {
             <div className="sticky top-0 z-10 border-b border-t bg-background" ref={tabRef}>
               <div className="relative mx-auto -mb-px h-16 max-w-screen-xl px-6">
                 <TabsList centered={false}>
-                  <TabsTrigger href={baseRoute} value="fundraiser" activeTab={activeTab}>
+                  <TabsTrigger href={baseRoute} active={activeTab === 'fundraiser'}>
                     Fundraiser
+                  </TabsTrigger>
+                  <TabsTrigger href={`${baseRoute}/finances`} active={activeTab === 'finances'}>
+                    Finances
                   </TabsTrigger>
                   <TabsTrigger
                     href={`${baseRoute}/updates`}
-                    value="updates"
-                    activeTab={activeTab}
+                    active={activeTab === 'updates'}
                     count={account.updates?.totalCount}
                   >
                     Updates
                   </TabsTrigger>
-                  <TabsTrigger
-                    href={`${baseRoute}/contributions`}
-                    value="contributions"
-                    activeTab={activeTab}
-                    count={account.contributionTransactions?.totalCount}
-                  >
-                    Contributions
-                  </TabsTrigger>
-                  <TabsTrigger
-                    href={`${baseRoute}/expenses`}
-                    value="expenses"
-                    activeTab={activeTab}
-                    count={account.expenses?.totalCount}
-                  >
-                    Expenses
-                  </TabsTrigger>
                 </TabsList>
               </div>
             </div>
-            {children}
+            <div className="flex-1">{children}</div>
 
             <ThemeColor color={fundraiser.primaryColor} />
           </React.Fragment>
