+++ components/crowdfunding-redesign/ProfileLayout.tsx
@@ -1,21 +1,22 @@
 import React from 'react';
-
-import { getDefaultProfileValues } from './helpers';
-
-import { ThemeColor } from './ThemeColor';
-import { profileWrapperQuery } from './queries';
 import { useQuery } from '@apollo/client';
-import { API_V2_CONTEXT } from '../../lib/graphql/helpers';
-import { useRouter } from 'next/router';
-import { Footer } from './Footer';
-import Page from '../Page';
 import { AnimatePresence } from 'framer-motion';
-import { Header } from './Header';
+import { useRouter } from 'next/router';
+
+import { API_V2_CONTEXT } from '../../lib/graphql/helpers';
+
+import Page from '../Page';
+
 import { Banner } from './Banner';
+import { Footer } from './Footer';
+import { Header } from './Header';
+import { getDefaultProfileValues } from './helpers';
+import { profileWrapperQuery } from './queries';
+import { ThemeColor } from './ThemeColor';
 
 export default function ProfileLayout({ activeTab = 'home', children, collapsed = false, getBreadcrumbs = undefined }) {
   const router = useRouter();
-  const { data, loading, variables, error } = useQuery(profileWrapperQuery, {
+  const { data } = useQuery(profileWrapperQuery, {
     variables: {
       collectiveSlug: router.query.collectiveSlug,
       accountSlug: router.query.accountSlug,
@@ -32,13 +33,12 @@ export default function ProfileLayout({ activeTab = 'home', children, collapsed 
 
   return (
     <Page withTopBar={false} showFooter={false} title={`[PREVIEW] ${collective?.name} | Open Collective`}>
-      {/* <Banner account={account || collective} isFundraiser={account?.type === 'EVENT' || account?.type === 'PROJECT'} /> */}
+      <Banner account={account || collective} isFundraiser={account?.type === 'EVENT' || account?.type === 'PROJECT'} />
 
       <AnimatePresence mode="wait" initial={false}>
         <div className="flex min-h-screen flex-col justify-between bg-primary-foreground/75 antialiased">
           <Header
             collective={collective}
-            account={account}
             activeTab={activeTab}
             profile={profile}
             collapsed={collapsed}
