+++ components/crowdfunding-redesign/Tabs.tsx
@@ -1,44 +1,26 @@
 import React from 'react';
-import * as TabsPrimitive from '@radix-ui/react-tabs';
+// eslint-disable-next-line no-restricted-imports
+import Link from 'next/link';
 
 import { cn } from '../../lib/utils';
 
-const Tabs = TabsPrimitive.Root;
+const TabsList = ({ className = '', centered = true, ...props }) => (
+  <div className={cn('flex h-full gap-8', centered && 'justify-center', className)} {...props} />
+);
 
-const TabsList = React.forwardRef<
-  React.ElementRef<typeof TabsPrimitive.List>,
-  React.ComponentPropsWithoutRef<typeof TabsPrimitive.List> & { centered?: boolean }
->(({ className, centered, ...props }, ref) => (
-  <TabsPrimitive.List
-    ref={ref}
-    className={cn('flex h-full gap-8', centered && 'justify-center', className)}
-    {...props}
-  />
-));
-TabsList.displayName = TabsPrimitive.List.displayName;
-
-const TabsTrigger = React.forwardRef<
-  React.ElementRef<typeof TabsPrimitive.Trigger>,
-  React.ComponentPropsWithoutRef<typeof TabsPrimitive.Trigger> & { count?: number }
->(({ className, children, count, ...props }, ref) => (
-  <TabsPrimitive.Trigger
-    ref={ref}
+const TabsTrigger = ({ className = '', children, href, active, count = undefined, ...props }) => (
+  <Link
+    href={href}
     className={cn(
       'relative flex h-full items-center gap-1.5 border-b-[3px] px-1 text-sm antialiased hover:text-primary',
-      'border-transparent font-medium transition-colors data-[state=active]:border-black',
+      'border-transparent font-medium transition-colors',
+      active && 'border-black',
       className,
     )}
     {...props}
   >
     {children} {count > 0 && <div className="relative -top-1 align-top text-xs text-primary">{count}</div>}
-  </TabsPrimitive.Trigger>
-));
-TabsTrigger.displayName = TabsPrimitive.Trigger.displayName;
+  </Link>
+);
 
-const TabsContent = React.forwardRef<
-  React.ElementRef<typeof TabsPrimitive.Content>,
-  React.ComponentPropsWithoutRef<typeof TabsPrimitive.Content>
->(({ className, ...props }, ref) => <TabsPrimitive.Content ref={ref} className={cn('', className)} {...props} />);
-TabsContent.displayName = TabsPrimitive.Content.displayName;
-
-export { Tabs, TabsList, TabsTrigger, TabsContent };
+export { TabsList, TabsTrigger };
