+++ components/crowdfunding-redesign/updates/SingleUpdate.tsx
@@ -2,6 +2,8 @@ import React from 'react';
 import { useQuery } from '@apollo/client';
 import { Separator } from '@radix-ui/react-dropdown-menu';
 import { Markup } from 'interweave';
+import { ArrowLeft } from 'lucide-react';
+import { useRouter } from 'next/router';
 
 import { API_V2_CONTEXT } from '../../../lib/graphql/helpers';
 
@@ -10,13 +12,11 @@ import EmojiReactionPicker from '../../conversations/EmojiReactionPicker';
 import EmojiReactions from '../../conversations/EmojiReactions';
 import Thread from '../../conversations/Thread';
 import { updatesViewQuery } from '../../dashboard/sections/updates/queries';
+import Link from '../../Link';
+import { Button } from '../../ui/Button';
 import { Skeleton } from '../../ui/Skeleton';
 
 import { UpdateHeader } from './UpdateHeader';
-import { useRouter } from 'next/router';
-import { Button } from '../../ui/Button';
-import Link from '../../Link';
-import { ArrowLeft } from 'lucide-react';
 
 export function SingleUpdate() {
   const router = useRouter();
