+++ lib/graphql/types/v2/gql.ts
@@ -47,7 +47,13 @@ const documents = {
     "\n  query IsUserFollowingConversation($id: String!) {\n    loggedInAccount {\n      id\n      slug\n      imageUrl\n      type\n      name\n      ... on Individual {\n        isFollowingConversation(id: $id)\n      }\n    }\n  }\n": types.IsUserFollowingConversationDocument,
     "\n  mutation EditComment($comment: CommentUpdateInput!) {\n    editComment(comment: $comment) {\n      id\n      ...CommentFields\n    }\n  }\n  \n": types.EditCommentDocument,
     "\n  mutation EditCrowdfundingRedesignSettings($account: AccountReferenceInput!, $value: JSON!) {\n    editAccountSetting(account: $account, key: \"crowdfundingRedesign\", value: $value) {\n      id\n      settings\n    }\n  }\n": types.EditCrowdfundingRedesignSettingsDocument,
-    "\n  query ProfilePage($slug: String!, $includeChildren: Boolean!) {\n    account(slug: $slug) {\n      id\n      name\n      slug\n      imageUrl\n      description\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n      socialLinks {\n        type\n        url\n      }\n\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n        nodes {\n          id\n          title\n          html\n          slug\n          summary\n          reactions\n          userReactions\n          userCanSeeUpdate\n          publishedAt\n          fromAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n          comments {\n            totalCount\n          }\n        }\n      }\n      expenses(limit: 20, direction: RECEIVED, status: PAID, includeChildrenExpenses: $includeChildren) {\n        totalCount\n        nodes {\n          id\n          description\n          type\n          status\n          createdAt\n          payee {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          account {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          amountV2 {\n            valueInCents\n            currency\n          }\n        }\n      }\n\n      contributionTransactions: transactions(\n        limit: 20\n        kind: [CONTRIBUTION, ADDED_FUNDS]\n        isRefund: false\n        type: CREDIT\n      ) {\n        totalCount\n        nodes {\n          id\n          description\n          createdAt\n          type\n          kind\n          fromAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          toAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          amount {\n            valueInCents\n            currency\n          }\n        }\n      }\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n      ... on AccountWithHost {\n        host {\n          id\n          slug\n          name\n          imageUrl\n        }\n      }\n\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n\n        financialContributors: contributors(roles: [BACKER], limit: 150) {\n          totalCount\n          nodes {\n            id\n            name\n            roles\n            isAdmin\n            isCore\n            isBacker\n            since\n            image\n            description\n            collectiveSlug\n            totalAmountDonated\n            type\n            publicMessage\n            isIncognito\n          }\n        }\n      }\n      childrenAccounts {\n        nodes {\n          id\n          name\n          description\n          type\n          slug\n          settings\n          currency\n          stats {\n            totalAmountReceived {\n              valueInCents\n              currency\n            }\n            yearlyBudget {\n              valueInCents\n              currency\n            }\n            contributorsCount\n          }\n        }\n      }\n\n      stats {\n        totalAmountReceived {\n          valueInCents\n          currency\n        }\n        yearlyBudget {\n          valueInCents\n          currency\n        }\n        contributorsCount\n      }\n    }\n  }\n": types.ProfilePageDocument,
+    "\n  query TransactionGroupDetails($slug: String, $groupId: String!) {\n    transactionGroup(account: { slug: $slug }, groupId: $groupId) {\n      id\n      totalAmount {\n        valueInCents\n        currency\n      }\n      account {\n        id\n        name\n        slug\n        imageUrl\n      }\n      createdAt\n      primaryTransaction {\n        kind\n        type\n        oppositeAccount {\n          id\n          name\n          slug\n          imageUrl\n          type\n        }\n      }\n      transactions {\n        id\n        kind\n        type\n        account {\n          id\n          name\n          slug\n          imageUrl\n          type\n        }\n        oppositeAccount {\n          id\n          name\n          slug\n          imageUrl\n          type\n        }\n        expense {\n          id\n          status\n          amountV2 {\n            valueInCents\n            currency\n          }\n        }\n        order {\n          id\n        }\n        amount {\n          valueInCents\n          currency\n        }\n      }\n    }\n  }\n": types.TransactionGroupDetailsDocument,
+    "\n  query TransactionGroups($slug: String!, $limit: Int) {\n    transactionGroups(account: { slug: $slug }, limit: $limit) {\n      totalCount\n      nodes {\n        id\n        totalAmount {\n          valueInCents\n          currency\n        }\n        createdAt\n        primaryTransaction {\n          id\n          kind\n          type\n          oppositeAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n        }\n      }\n    }\n  }\n": types.TransactionGroupsDocument,
+    "\n  query Finances(\n    $slug: String!\n    $dateFrom: DateTime\n    $dateTo: DateTime\n    $compareFrom: DateTime\n    $compareTo: DateTime\n    $includeComparison: Boolean!\n    $includeBalance: Boolean!\n    $includeSpent: Boolean!\n    $includeReceived: Boolean!\n    $includeBalanceTimeseries: Boolean!\n    $includeReceivedTimeseries: Boolean!\n    $timeUnit: TimeUnit\n    $includeChildren: Boolean!\n  ) {\n    account(slug: $slug) {\n      id\n      name\n      type\n\n      totalBalance: stats {\n        id\n        current: balance(includeChildren: $includeChildren, dateTo: $dateTo) {\n          currency\n          valueInCents\n        }\n        comparison: balance(includeChildren: $includeChildren, dateTo: $compareTo) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n      }\n\n      balanceTimeseries: stats @include(if: $includeBalanceTimeseries) {\n        id\n        current: balanceTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          timeUnit: $timeUnit\n        ) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n        comparison: balanceTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          timeUnit: $timeUnit\n        ) @include(if: $includeComparison) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n      }\n\n      totalSpent: stats {\n        id\n        current: totalAmountSpent(includeChildren: $includeChildren, dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n          currency\n          valueInCents\n        }\n        comparison: totalAmountSpent(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n      }\n      totalReceived: stats {\n        id\n        current: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          net: true\n        ) {\n          currency\n          valueInCents\n        }\n        comparison: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n      }\n\n      receivedTimeseries: stats @include(if: $includeReceivedTimeseries) {\n        id\n        current: totalAmountReceivedTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          timeUnit: $timeUnit\n          net: true\n        ) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n        comparison: totalAmountReceivedTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          timeUnit: $timeUnit\n          net: true\n        ) @include(if: $includeComparison) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n      }\n\n      ...ProfileMetrics\n\n      childrenAccounts {\n        totalCount\n        nodes {\n          id\n          ...ProfileMetrics\n        }\n      }\n    }\n  }\n\n  fragment ProfileMetrics on Account {\n    ...AccountHoverCardFields\n    balance: stats @include(if: $includeBalance) {\n      id\n      current: balance(dateTo: $dateTo) {\n        currency\n        valueInCents\n      }\n      # comparison: balance(dateTo: $compareTo) @include(if: $includeComparison) {\n      #   currency\n      #   valueInCents\n      # }\n    }\n    spent: stats @include(if: $includeSpent) {\n      id\n      current: totalAmountSpent(dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n        currency\n        valueInCents\n      }\n      # comparison: totalAmountSpent(dateFrom: $compareFrom, dateTo: $compareTo, net: true)\n      #   @include(if: $includeComparison) {\n      #   currency\n      #   valueInCents\n      # }\n    }\n    received: stats @include(if: $includeReceived) {\n      id\n      current: totalAmountReceived(dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n        currency\n        valueInCents\n      }\n      # comparison: totalAmountReceived(dateFrom: $compareFrom, dateTo: $compareTo, net: true)\n      #   @include(if: $includeComparison) {\n      #   currency\n      #   valueInCents\n      # }\n    }\n  }\n  \n": types.FinancesDocument,
+    "\n  query Updates($slug: String!) {\n    account(slug: $slug) {\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n        nodes {\n          id\n          title\n          html\n          slug\n          summary\n          reactions\n          userReactions\n          userCanSeeUpdate\n          publishedAt\n          fromAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n          comments {\n            totalCount\n          }\n        }\n      }\n    }\n  }\n": types.UpdatesDocument,
+    "\n  query ProfilePage($slug: String!) {\n    account(slug: $slug) {\n      id\n      name\n      slug\n      imageUrl\n      description\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n      socialLinks {\n        type\n        url\n      }\n\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n        nodes {\n          id\n          title\n          html\n          slug\n          summary\n          reactions\n          userReactions\n          userCanSeeUpdate\n          publishedAt\n          fromAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n          comments {\n            totalCount\n          }\n        }\n      }\n      expenses(limit: 20, direction: RECEIVED, status: PAID) {\n        totalCount\n        nodes {\n          id\n          description\n          type\n          status\n          createdAt\n          payee {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          account {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          amountV2 {\n            valueInCents\n            currency\n          }\n        }\n      }\n\n      contributionTransactions: transactions(\n        limit: 20\n        kind: [CONTRIBUTION, ADDED_FUNDS]\n        isRefund: false\n        type: CREDIT\n      ) {\n        totalCount\n        nodes {\n          id\n          description\n          createdAt\n          type\n          kind\n          fromAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          toAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          amount {\n            valueInCents\n            currency\n          }\n        }\n      }\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n      ... on AccountWithHost {\n        host {\n          id\n          slug\n          name\n          imageUrl\n        }\n      }\n\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n\n        financialContributors: contributors(roles: [BACKER], limit: 150) {\n          totalCount\n          nodes {\n            id\n            name\n            roles\n            isAdmin\n            isCore\n            isBacker\n            since\n            image\n            description\n            collectiveSlug\n            totalAmountDonated\n            type\n            publicMessage\n            isIncognito\n          }\n        }\n      }\n      childrenAccounts {\n        nodes {\n          id\n          name\n          description\n          type\n          slug\n          settings\n          currency\n          stats {\n            totalAmountReceived {\n              valueInCents\n              currency\n            }\n            yearlyBudget {\n              valueInCents\n              currency\n            }\n            contributorsCount\n          }\n        }\n      }\n\n      stats {\n        totalAmountReceived {\n          valueInCents\n          currency\n        }\n        yearlyBudget {\n          valueInCents\n          currency\n        }\n        contributorsCount\n      }\n    }\n  }\n": types.ProfilePageDocument,
+    "\n  query ProfileLayout($collectiveSlug: String!, $accountSlug: String, $includeAccount: Boolean!) {\n    account: account(slug: $accountSlug) @include(if: $includeAccount) {\n      id\n      name\n      slug\n      settings\n      type\n      description\n      longDescription\n      backgroundImageUrl\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n        }\n      }\n    }\n    collective: account(slug: $collectiveSlug) {\n      id\n      name\n      slug\n      imageUrl\n      description\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n      socialLinks {\n        type\n        url\n      }\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n      }\n\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n      }\n      expenses(limit: 0, direction: RECEIVED, status: PAID, includeChildrenExpenses: true) {\n        totalCount\n      }\n\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n      ... on AccountWithHost {\n        host {\n          id\n          slug\n          name\n          imageUrl\n        }\n      }\n      childrenAccounts {\n        nodes {\n          id\n          name\n          description\n          type\n          slug\n          settings\n          currency\n          stats {\n            totalAmountReceived {\n              valueInCents\n              currency\n            }\n            yearlyBudget {\n              valueInCents\n              currency\n            }\n            contributorsCount\n          }\n        }\n      }\n    }\n  }\n": types.ProfileLayoutDocument,
+    "\n  query ContributePage($slug: String!) {\n    account(slug: $slug) {\n      id\n      name\n      slug\n      imageUrl\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n\n      stats {\n        totalAmountReceived {\n          valueInCents\n          currency\n        }\n        yearlyBudget {\n          valueInCents\n          currency\n        }\n        contributorsCount\n      }\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n\n        financialContributors: contributors(roles: [BACKER], limit: 150) {\n          totalCount\n          nodes {\n            id\n            name\n            roles\n            isAdmin\n            isCore\n            isBacker\n            since\n            image\n            description\n            collectiveSlug\n            totalAmountDonated\n            type\n            publicMessage\n            isIncognito\n          }\n        }\n      }\n\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n    }\n    projects: accounts(\n      parent: { slug: $slug }\n      type: [PROJECT]\n      isActive: true\n      orderBy: { field: BALANCE, direction: DESC }\n    ) {\n      totalCount\n      nodes {\n        id\n        name\n        description\n        type\n        slug\n        settings\n        currency\n        stats {\n          totalAmountReceived {\n            valueInCents\n            currency\n          }\n          yearlyBudget {\n            valueInCents\n            currency\n          }\n          contributorsCount\n        }\n      }\n    }\n    events: accounts(\n      parent: { slug: $slug }\n      type: [EVENT]\n      isActive: true\n      orderBy: { field: BALANCE, direction: DESC }\n    ) {\n      totalCount\n      nodes {\n        id\n        name\n        description\n        type\n        slug\n        settings\n        currency\n        stats {\n          totalAmountReceived {\n            valueInCents\n            currency\n          }\n          yearlyBudget {\n            valueInCents\n            currency\n          }\n          contributorsCount\n        }\n      }\n    }\n  }\n": types.ContributePageDocument,
     "\n  query Account($slug: String!) {\n    account(slug: $slug) {\n      id\n      ...AccountHoverCardFields\n    }\n  }\n  \n": types.AccountDocument,
     "\n  query AccountFilterSearch($searchTerm: String) {\n    accounts(searchTerm: $searchTerm) {\n      nodes {\n        id\n        ...AccountHoverCardFields\n      }\n    }\n  }\n  \n": types.AccountFilterSearchDocument,
     "\n  query AccountingCategories($hostSlug: String) {\n    host(slug: $hostSlug) {\n      id\n      accountingCategories {\n        nodes {\n          id\n          code\n          name\n          kind\n          appliesTo\n        }\n      }\n    }\n  }\n": types.AccountingCategoriesDocument,
@@ -399,7 +405,31 @@ export function graphql(source: "\n  mutation EditCrowdfundingRedesignSettings($
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query ProfilePage($slug: String!, $includeChildren: Boolean!) {\n    account(slug: $slug) {\n      id\n      name\n      slug\n      imageUrl\n      description\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n      socialLinks {\n        type\n        url\n      }\n\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n        nodes {\n          id\n          title\n          html\n          slug\n          summary\n          reactions\n          userReactions\n          userCanSeeUpdate\n          publishedAt\n          fromAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n          comments {\n            totalCount\n          }\n        }\n      }\n      expenses(limit: 20, direction: RECEIVED, status: PAID, includeChildrenExpenses: $includeChildren) {\n        totalCount\n        nodes {\n          id\n          description\n          type\n          status\n          createdAt\n          payee {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          account {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          amountV2 {\n            valueInCents\n            currency\n          }\n        }\n      }\n\n      contributionTransactions: transactions(\n        limit: 20\n        kind: [CONTRIBUTION, ADDED_FUNDS]\n        isRefund: false\n        type: CREDIT\n      ) {\n        totalCount\n        nodes {\n          id\n          description\n          createdAt\n          type\n          kind\n          fromAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          toAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          amount {\n            valueInCents\n            currency\n          }\n        }\n      }\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n      ... on AccountWithHost {\n        host {\n          id\n          slug\n          name\n          imageUrl\n        }\n      }\n\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n\n        financialContributors: contributors(roles: [BACKER], limit: 150) {\n          totalCount\n          nodes {\n            id\n            name\n            roles\n            isAdmin\n            isCore\n            isBacker\n            since\n            image\n            description\n            collectiveSlug\n            totalAmountDonated\n            type\n            publicMessage\n            isIncognito\n          }\n        }\n      }\n      childrenAccounts {\n        nodes {\n          id\n          name\n          description\n          type\n          slug\n          settings\n          currency\n          stats {\n            totalAmountReceived {\n              valueInCents\n              currency\n            }\n            yearlyBudget {\n              valueInCents\n              currency\n            }\n            contributorsCount\n          }\n        }\n      }\n\n      stats {\n        totalAmountReceived {\n          valueInCents\n          currency\n        }\n        yearlyBudget {\n          valueInCents\n          currency\n        }\n        contributorsCount\n      }\n    }\n  }\n"): (typeof documents)["\n  query ProfilePage($slug: String!, $includeChildren: Boolean!) {\n    account(slug: $slug) {\n      id\n      name\n      slug\n      imageUrl\n      description\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n      socialLinks {\n        type\n        url\n      }\n\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n        nodes {\n          id\n          title\n          html\n          slug\n          summary\n          reactions\n          userReactions\n          userCanSeeUpdate\n          publishedAt\n          fromAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n          comments {\n            totalCount\n          }\n        }\n      }\n      expenses(limit: 20, direction: RECEIVED, status: PAID, includeChildrenExpenses: $includeChildren) {\n        totalCount\n        nodes {\n          id\n          description\n          type\n          status\n          createdAt\n          payee {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          account {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          amountV2 {\n            valueInCents\n            currency\n          }\n        }\n      }\n\n      contributionTransactions: transactions(\n        limit: 20\n        kind: [CONTRIBUTION, ADDED_FUNDS]\n        isRefund: false\n        type: CREDIT\n      ) {\n        totalCount\n        nodes {\n          id\n          description\n          createdAt\n          type\n          kind\n          fromAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          toAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          amount {\n            valueInCents\n            currency\n          }\n        }\n      }\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n      ... on AccountWithHost {\n        host {\n          id\n          slug\n          name\n          imageUrl\n        }\n      }\n\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n\n        financialContributors: contributors(roles: [BACKER], limit: 150) {\n          totalCount\n          nodes {\n            id\n            name\n            roles\n            isAdmin\n            isCore\n            isBacker\n            since\n            image\n            description\n            collectiveSlug\n            totalAmountDonated\n            type\n            publicMessage\n            isIncognito\n          }\n        }\n      }\n      childrenAccounts {\n        nodes {\n          id\n          name\n          description\n          type\n          slug\n          settings\n          currency\n          stats {\n            totalAmountReceived {\n              valueInCents\n              currency\n            }\n            yearlyBudget {\n              valueInCents\n              currency\n            }\n            contributorsCount\n          }\n        }\n      }\n\n      stats {\n        totalAmountReceived {\n          valueInCents\n          currency\n        }\n        yearlyBudget {\n          valueInCents\n          currency\n        }\n        contributorsCount\n      }\n    }\n  }\n"];
+export function graphql(source: "\n  query TransactionGroupDetails($slug: String, $groupId: String!) {\n    transactionGroup(account: { slug: $slug }, groupId: $groupId) {\n      id\n      totalAmount {\n        valueInCents\n        currency\n      }\n      account {\n        id\n        name\n        slug\n        imageUrl\n      }\n      createdAt\n      primaryTransaction {\n        kind\n        type\n        oppositeAccount {\n          id\n          name\n          slug\n          imageUrl\n          type\n        }\n      }\n      transactions {\n        id\n        kind\n        type\n        account {\n          id\n          name\n          slug\n          imageUrl\n          type\n        }\n        oppositeAccount {\n          id\n          name\n          slug\n          imageUrl\n          type\n        }\n        expense {\n          id\n          status\n          amountV2 {\n            valueInCents\n            currency\n          }\n        }\n        order {\n          id\n        }\n        amount {\n          valueInCents\n          currency\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query TransactionGroupDetails($slug: String, $groupId: String!) {\n    transactionGroup(account: { slug: $slug }, groupId: $groupId) {\n      id\n      totalAmount {\n        valueInCents\n        currency\n      }\n      account {\n        id\n        name\n        slug\n        imageUrl\n      }\n      createdAt\n      primaryTransaction {\n        kind\n        type\n        oppositeAccount {\n          id\n          name\n          slug\n          imageUrl\n          type\n        }\n      }\n      transactions {\n        id\n        kind\n        type\n        account {\n          id\n          name\n          slug\n          imageUrl\n          type\n        }\n        oppositeAccount {\n          id\n          name\n          slug\n          imageUrl\n          type\n        }\n        expense {\n          id\n          status\n          amountV2 {\n            valueInCents\n            currency\n          }\n        }\n        order {\n          id\n        }\n        amount {\n          valueInCents\n          currency\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query TransactionGroups($slug: String!, $limit: Int) {\n    transactionGroups(account: { slug: $slug }, limit: $limit) {\n      totalCount\n      nodes {\n        id\n        totalAmount {\n          valueInCents\n          currency\n        }\n        createdAt\n        primaryTransaction {\n          id\n          kind\n          type\n          oppositeAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query TransactionGroups($slug: String!, $limit: Int) {\n    transactionGroups(account: { slug: $slug }, limit: $limit) {\n      totalCount\n      nodes {\n        id\n        totalAmount {\n          valueInCents\n          currency\n        }\n        createdAt\n        primaryTransaction {\n          id\n          kind\n          type\n          oppositeAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query Finances(\n    $slug: String!\n    $dateFrom: DateTime\n    $dateTo: DateTime\n    $compareFrom: DateTime\n    $compareTo: DateTime\n    $includeComparison: Boolean!\n    $includeBalance: Boolean!\n    $includeSpent: Boolean!\n    $includeReceived: Boolean!\n    $includeBalanceTimeseries: Boolean!\n    $includeReceivedTimeseries: Boolean!\n    $timeUnit: TimeUnit\n    $includeChildren: Boolean!\n  ) {\n    account(slug: $slug) {\n      id\n      name\n      type\n\n      totalBalance: stats {\n        id\n        current: balance(includeChildren: $includeChildren, dateTo: $dateTo) {\n          currency\n          valueInCents\n        }\n        comparison: balance(includeChildren: $includeChildren, dateTo: $compareTo) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n      }\n\n      balanceTimeseries: stats @include(if: $includeBalanceTimeseries) {\n        id\n        current: balanceTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          timeUnit: $timeUnit\n        ) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n        comparison: balanceTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          timeUnit: $timeUnit\n        ) @include(if: $includeComparison) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n      }\n\n      totalSpent: stats {\n        id\n        current: totalAmountSpent(includeChildren: $includeChildren, dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n          currency\n          valueInCents\n        }\n        comparison: totalAmountSpent(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n      }\n      totalReceived: stats {\n        id\n        current: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          net: true\n        ) {\n          currency\n          valueInCents\n        }\n        comparison: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n      }\n\n      receivedTimeseries: stats @include(if: $includeReceivedTimeseries) {\n        id\n        current: totalAmountReceivedTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          timeUnit: $timeUnit\n          net: true\n        ) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n        comparison: totalAmountReceivedTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          timeUnit: $timeUnit\n          net: true\n        ) @include(if: $includeComparison) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n      }\n\n      ...ProfileMetrics\n\n      childrenAccounts {\n        totalCount\n        nodes {\n          id\n          ...ProfileMetrics\n        }\n      }\n    }\n  }\n\n  fragment ProfileMetrics on Account {\n    ...AccountHoverCardFields\n    balance: stats @include(if: $includeBalance) {\n      id\n      current: balance(dateTo: $dateTo) {\n        currency\n        valueInCents\n      }\n      # comparison: balance(dateTo: $compareTo) @include(if: $includeComparison) {\n      #   currency\n      #   valueInCents\n      # }\n    }\n    spent: stats @include(if: $includeSpent) {\n      id\n      current: totalAmountSpent(dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n        currency\n        valueInCents\n      }\n      # comparison: totalAmountSpent(dateFrom: $compareFrom, dateTo: $compareTo, net: true)\n      #   @include(if: $includeComparison) {\n      #   currency\n      #   valueInCents\n      # }\n    }\n    received: stats @include(if: $includeReceived) {\n      id\n      current: totalAmountReceived(dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n        currency\n        valueInCents\n      }\n      # comparison: totalAmountReceived(dateFrom: $compareFrom, dateTo: $compareTo, net: true)\n      #   @include(if: $includeComparison) {\n      #   currency\n      #   valueInCents\n      # }\n    }\n  }\n  \n"): (typeof documents)["\n  query Finances(\n    $slug: String!\n    $dateFrom: DateTime\n    $dateTo: DateTime\n    $compareFrom: DateTime\n    $compareTo: DateTime\n    $includeComparison: Boolean!\n    $includeBalance: Boolean!\n    $includeSpent: Boolean!\n    $includeReceived: Boolean!\n    $includeBalanceTimeseries: Boolean!\n    $includeReceivedTimeseries: Boolean!\n    $timeUnit: TimeUnit\n    $includeChildren: Boolean!\n  ) {\n    account(slug: $slug) {\n      id\n      name\n      type\n\n      totalBalance: stats {\n        id\n        current: balance(includeChildren: $includeChildren, dateTo: $dateTo) {\n          currency\n          valueInCents\n        }\n        comparison: balance(includeChildren: $includeChildren, dateTo: $compareTo) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n      }\n\n      balanceTimeseries: stats @include(if: $includeBalanceTimeseries) {\n        id\n        current: balanceTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          timeUnit: $timeUnit\n        ) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n        comparison: balanceTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          timeUnit: $timeUnit\n        ) @include(if: $includeComparison) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n      }\n\n      totalSpent: stats {\n        id\n        current: totalAmountSpent(includeChildren: $includeChildren, dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n          currency\n          valueInCents\n        }\n        comparison: totalAmountSpent(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n      }\n      totalReceived: stats {\n        id\n        current: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          net: true\n        ) {\n          currency\n          valueInCents\n        }\n        comparison: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n      }\n\n      receivedTimeseries: stats @include(if: $includeReceivedTimeseries) {\n        id\n        current: totalAmountReceivedTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          timeUnit: $timeUnit\n          net: true\n        ) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n        comparison: totalAmountReceivedTimeSeries(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          timeUnit: $timeUnit\n          net: true\n        ) @include(if: $includeComparison) {\n          dateTo\n          dateFrom\n          timeUnit\n          nodes {\n            date\n            amount {\n              currency\n              value\n            }\n          }\n        }\n      }\n\n      ...ProfileMetrics\n\n      childrenAccounts {\n        totalCount\n        nodes {\n          id\n          ...ProfileMetrics\n        }\n      }\n    }\n  }\n\n  fragment ProfileMetrics on Account {\n    ...AccountHoverCardFields\n    balance: stats @include(if: $includeBalance) {\n      id\n      current: balance(dateTo: $dateTo) {\n        currency\n        valueInCents\n      }\n      # comparison: balance(dateTo: $compareTo) @include(if: $includeComparison) {\n      #   currency\n      #   valueInCents\n      # }\n    }\n    spent: stats @include(if: $includeSpent) {\n      id\n      current: totalAmountSpent(dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n        currency\n        valueInCents\n      }\n      # comparison: totalAmountSpent(dateFrom: $compareFrom, dateTo: $compareTo, net: true)\n      #   @include(if: $includeComparison) {\n      #   currency\n      #   valueInCents\n      # }\n    }\n    received: stats @include(if: $includeReceived) {\n      id\n      current: totalAmountReceived(dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n        currency\n        valueInCents\n      }\n      # comparison: totalAmountReceived(dateFrom: $compareFrom, dateTo: $compareTo, net: true)\n      #   @include(if: $includeComparison) {\n      #   currency\n      #   valueInCents\n      # }\n    }\n  }\n  \n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query Updates($slug: String!) {\n    account(slug: $slug) {\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n        nodes {\n          id\n          title\n          html\n          slug\n          summary\n          reactions\n          userReactions\n          userCanSeeUpdate\n          publishedAt\n          fromAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n          comments {\n            totalCount\n          }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query Updates($slug: String!) {\n    account(slug: $slug) {\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n        nodes {\n          id\n          title\n          html\n          slug\n          summary\n          reactions\n          userReactions\n          userCanSeeUpdate\n          publishedAt\n          fromAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n          comments {\n            totalCount\n          }\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query ProfilePage($slug: String!) {\n    account(slug: $slug) {\n      id\n      name\n      slug\n      imageUrl\n      description\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n      socialLinks {\n        type\n        url\n      }\n\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n        nodes {\n          id\n          title\n          html\n          slug\n          summary\n          reactions\n          userReactions\n          userCanSeeUpdate\n          publishedAt\n          fromAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n          comments {\n            totalCount\n          }\n        }\n      }\n      expenses(limit: 20, direction: RECEIVED, status: PAID) {\n        totalCount\n        nodes {\n          id\n          description\n          type\n          status\n          createdAt\n          payee {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          account {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          amountV2 {\n            valueInCents\n            currency\n          }\n        }\n      }\n\n      contributionTransactions: transactions(\n        limit: 20\n        kind: [CONTRIBUTION, ADDED_FUNDS]\n        isRefund: false\n        type: CREDIT\n      ) {\n        totalCount\n        nodes {\n          id\n          description\n          createdAt\n          type\n          kind\n          fromAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          toAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          amount {\n            valueInCents\n            currency\n          }\n        }\n      }\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n      ... on AccountWithHost {\n        host {\n          id\n          slug\n          name\n          imageUrl\n        }\n      }\n\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n\n        financialContributors: contributors(roles: [BACKER], limit: 150) {\n          totalCount\n          nodes {\n            id\n            name\n            roles\n            isAdmin\n            isCore\n            isBacker\n            since\n            image\n            description\n            collectiveSlug\n            totalAmountDonated\n            type\n            publicMessage\n            isIncognito\n          }\n        }\n      }\n      childrenAccounts {\n        nodes {\n          id\n          name\n          description\n          type\n          slug\n          settings\n          currency\n          stats {\n            totalAmountReceived {\n              valueInCents\n              currency\n            }\n            yearlyBudget {\n              valueInCents\n              currency\n            }\n            contributorsCount\n          }\n        }\n      }\n\n      stats {\n        totalAmountReceived {\n          valueInCents\n          currency\n        }\n        yearlyBudget {\n          valueInCents\n          currency\n        }\n        contributorsCount\n      }\n    }\n  }\n"): (typeof documents)["\n  query ProfilePage($slug: String!) {\n    account(slug: $slug) {\n      id\n      name\n      slug\n      imageUrl\n      description\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n      socialLinks {\n        type\n        url\n      }\n\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n        nodes {\n          id\n          title\n          html\n          slug\n          summary\n          reactions\n          userReactions\n          userCanSeeUpdate\n          publishedAt\n          fromAccount {\n            id\n            name\n            slug\n            imageUrl\n            type\n          }\n          comments {\n            totalCount\n          }\n        }\n      }\n      expenses(limit: 20, direction: RECEIVED, status: PAID) {\n        totalCount\n        nodes {\n          id\n          description\n          type\n          status\n          createdAt\n          payee {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          account {\n            id\n            name\n            type\n            imageUrl\n            slug\n          }\n          amountV2 {\n            valueInCents\n            currency\n          }\n        }\n      }\n\n      contributionTransactions: transactions(\n        limit: 20\n        kind: [CONTRIBUTION, ADDED_FUNDS]\n        isRefund: false\n        type: CREDIT\n      ) {\n        totalCount\n        nodes {\n          id\n          description\n          createdAt\n          type\n          kind\n          fromAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          toAccount {\n            id\n            name\n            type\n            slug\n            imageUrl\n          }\n          amount {\n            valueInCents\n            currency\n          }\n        }\n      }\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n      ... on AccountWithHost {\n        host {\n          id\n          slug\n          name\n          imageUrl\n        }\n      }\n\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n\n        financialContributors: contributors(roles: [BACKER], limit: 150) {\n          totalCount\n          nodes {\n            id\n            name\n            roles\n            isAdmin\n            isCore\n            isBacker\n            since\n            image\n            description\n            collectiveSlug\n            totalAmountDonated\n            type\n            publicMessage\n            isIncognito\n          }\n        }\n      }\n      childrenAccounts {\n        nodes {\n          id\n          name\n          description\n          type\n          slug\n          settings\n          currency\n          stats {\n            totalAmountReceived {\n              valueInCents\n              currency\n            }\n            yearlyBudget {\n              valueInCents\n              currency\n            }\n            contributorsCount\n          }\n        }\n      }\n\n      stats {\n        totalAmountReceived {\n          valueInCents\n          currency\n        }\n        yearlyBudget {\n          valueInCents\n          currency\n        }\n        contributorsCount\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query ProfileLayout($collectiveSlug: String!, $accountSlug: String, $includeAccount: Boolean!) {\n    account: account(slug: $accountSlug) @include(if: $includeAccount) {\n      id\n      name\n      slug\n      settings\n      type\n      description\n      longDescription\n      backgroundImageUrl\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n        }\n      }\n    }\n    collective: account(slug: $collectiveSlug) {\n      id\n      name\n      slug\n      imageUrl\n      description\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n      socialLinks {\n        type\n        url\n      }\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n      }\n\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n      }\n      expenses(limit: 0, direction: RECEIVED, status: PAID, includeChildrenExpenses: true) {\n        totalCount\n      }\n\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n      ... on AccountWithHost {\n        host {\n          id\n          slug\n          name\n          imageUrl\n        }\n      }\n      childrenAccounts {\n        nodes {\n          id\n          name\n          description\n          type\n          slug\n          settings\n          currency\n          stats {\n            totalAmountReceived {\n              valueInCents\n              currency\n            }\n            yearlyBudget {\n              valueInCents\n              currency\n            }\n            contributorsCount\n          }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query ProfileLayout($collectiveSlug: String!, $accountSlug: String, $includeAccount: Boolean!) {\n    account: account(slug: $accountSlug) @include(if: $includeAccount) {\n      id\n      name\n      slug\n      settings\n      type\n      description\n      longDescription\n      backgroundImageUrl\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n        }\n      }\n    }\n    collective: account(slug: $collectiveSlug) {\n      id\n      name\n      slug\n      imageUrl\n      description\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n      socialLinks {\n        type\n        url\n      }\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n      }\n\n      updates(onlyPublishedUpdates: true) {\n        totalCount\n      }\n      expenses(limit: 0, direction: RECEIVED, status: PAID, includeChildrenExpenses: true) {\n        totalCount\n      }\n\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n      ... on AccountWithHost {\n        host {\n          id\n          slug\n          name\n          imageUrl\n        }\n      }\n      childrenAccounts {\n        nodes {\n          id\n          name\n          description\n          type\n          slug\n          settings\n          currency\n          stats {\n            totalAmountReceived {\n              valueInCents\n              currency\n            }\n            yearlyBudget {\n              valueInCents\n              currency\n            }\n            contributorsCount\n          }\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query ContributePage($slug: String!) {\n    account(slug: $slug) {\n      id\n      name\n      slug\n      imageUrl\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n\n      stats {\n        totalAmountReceived {\n          valueInCents\n          currency\n        }\n        yearlyBudget {\n          valueInCents\n          currency\n        }\n        contributorsCount\n      }\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n\n        financialContributors: contributors(roles: [BACKER], limit: 150) {\n          totalCount\n          nodes {\n            id\n            name\n            roles\n            isAdmin\n            isCore\n            isBacker\n            since\n            image\n            description\n            collectiveSlug\n            totalAmountDonated\n            type\n            publicMessage\n            isIncognito\n          }\n        }\n      }\n\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n    }\n    projects: accounts(\n      parent: { slug: $slug }\n      type: [PROJECT]\n      isActive: true\n      orderBy: { field: BALANCE, direction: DESC }\n    ) {\n      totalCount\n      nodes {\n        id\n        name\n        description\n        type\n        slug\n        settings\n        currency\n        stats {\n          totalAmountReceived {\n            valueInCents\n            currency\n          }\n          yearlyBudget {\n            valueInCents\n            currency\n          }\n          contributorsCount\n        }\n      }\n    }\n    events: accounts(\n      parent: { slug: $slug }\n      type: [EVENT]\n      isActive: true\n      orderBy: { field: BALANCE, direction: DESC }\n    ) {\n      totalCount\n      nodes {\n        id\n        name\n        description\n        type\n        slug\n        settings\n        currency\n        stats {\n          totalAmountReceived {\n            valueInCents\n            currency\n          }\n          yearlyBudget {\n            valueInCents\n            currency\n          }\n          contributorsCount\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query ContributePage($slug: String!) {\n    account(slug: $slug) {\n      id\n      name\n      slug\n      imageUrl\n      longDescription\n      backgroundImageUrl\n      settings\n      type\n      currency\n      isActive\n\n      stats {\n        totalAmountReceived {\n          valueInCents\n          currency\n        }\n        yearlyBudget {\n          valueInCents\n          currency\n        }\n        contributorsCount\n      }\n      ... on AccountWithContributions {\n        tiers {\n          totalCount\n          nodes {\n            id\n            name\n            slug\n            type\n            description\n            button\n            amount {\n              valueInCents\n              currency\n            }\n            currency\n            minimumAmount {\n              valueInCents\n              currency\n            }\n            endsAt\n            amountType\n            frequency\n            availableQuantity\n          }\n        }\n\n        financialContributors: contributors(roles: [BACKER], limit: 150) {\n          totalCount\n          nodes {\n            id\n            name\n            roles\n            isAdmin\n            isCore\n            isBacker\n            since\n            image\n            description\n            collectiveSlug\n            totalAmountDonated\n            type\n            publicMessage\n            isIncognito\n          }\n        }\n      }\n\n      ... on AccountWithParent {\n        parent {\n          id\n          slug\n          name\n          settings\n          imageUrl\n        }\n      }\n    }\n    projects: accounts(\n      parent: { slug: $slug }\n      type: [PROJECT]\n      isActive: true\n      orderBy: { field: BALANCE, direction: DESC }\n    ) {\n      totalCount\n      nodes {\n        id\n        name\n        description\n        type\n        slug\n        settings\n        currency\n        stats {\n          totalAmountReceived {\n            valueInCents\n            currency\n          }\n          yearlyBudget {\n            valueInCents\n            currency\n          }\n          contributorsCount\n        }\n      }\n    }\n    events: accounts(\n      parent: { slug: $slug }\n      type: [EVENT]\n      isActive: true\n      orderBy: { field: BALANCE, direction: DESC }\n    ) {\n      totalCount\n      nodes {\n        id\n        name\n        description\n        type\n        slug\n        settings\n        currency\n        stats {\n          totalAmountReceived {\n            valueInCents\n            currency\n          }\n          yearlyBudget {\n            valueInCents\n            currency\n          }\n          contributorsCount\n        }\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
