--- pages/preview/[collectiveSlug]/about.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import ProfileLayout from '../../../components/crowdfunding-redesign/ProfileLayout';
-
-export default function About() {
-  return (
-    <ProfileLayout activeTab="about">
-      <div className="flex-1">
-        <h1>About</h1>
-      </div>
-    </ProfileLayout>
-  );
-}
