+++ pages/preview/[collectiveSlug]/events/[accountSlug]/index.tsx
@@ -3,6 +3,8 @@ import React from 'react';
 import { FundraiserLayout } from '../../../../../components/crowdfunding-redesign/FundraiserLayout';
 import { FundraisingPage } from '../../../../../components/crowdfunding-redesign/FundraisingPage';
 
+// next.js export
+// ts-unused-exports:disable-next-line
 export default function EventFundraisingPage() {
   return (
     <FundraiserLayout activeTab="fundraiser">
