+++ pages/preview/[collectiveSlug]/events/[accountSlug]/updates.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import { FundraisingPage } from '../../../../../components/crowdfunding-redesign/FundraisingPage';
-import { useRouter } from 'next/router';
+
 import { FundraiserLayout } from '../../../../../components/crowdfunding-redesign/FundraiserLayout';
 import { UpdatesList } from '../../../../../components/crowdfunding-redesign/updates/UpdatesList';
 
+// next.js export
+// ts-unused-exports:disable-next-line
 export default function EventFundraisingPage() {
   return (
     <FundraiserLayout activeTab="updates">
