--- pages/preview/[collectiveSlug]/expenses.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import ProfileLayout from '../../../components/crowdfunding-redesign/ProfileLayout';
-
-export default function Expenses() {
-  return (
-    <ProfileLayout activeTab="expenses">
-      <div className="flex-1">
-        <h1>Expenses</h1>
-      </div>
-    </ProfileLayout>
-  );
-}
