--- pages/preview/[collectiveSlug]/path-test.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import { CrowdfundingPrototypePage } from '../../../components/crowdfunding-redesign/CrowdfundingPrototypePage';
-
-// next.js export
-// ts-unused-exports:disable-next-line
-export default function CrowdfundingPrototype() {
-  return <CrowdfundingPrototypePage />;
-}
