+++ pages/preview/[collectiveSlug]/updates/index.tsx
@@ -1,7 +1,10 @@
 import React from 'react';
+
 import ProfileLayout from '../../../../components/crowdfunding-redesign/ProfileLayout';
 import { UpdatesList } from '../../../../components/crowdfunding-redesign/updates/UpdatesList';
 
+// next.js export
+// ts-unused-exports:disable-next-line
 export default function Updates() {
   return (
     <ProfileLayout activeTab="updates">
